+++ docs/Whats-New.md
@@ -447,6 +447,7 @@ New:
 - Shield respawn animation and weapon (by Ollerus)
 - Customize the chained damage of the wall (by TaranDahl)
 - Allow the aircraft to enter area guard mission and not crash immediately without any airport (by CrimRecya)
+- Unlimbo Detonate warhead (by FlyStar)
 
 Vanilla fixes:
 - Fixed sidebar not updating queued unit numbers when adding or removing units when the production is on hold (by CrimRecya)